requirements.txt
json
configparser
argparse
colorama
pandas
base64
tencentcloud-sdk-python